Intial pass at "3-tier" architecture support.
The idea here is that condor_submit and the schedd live on different
machines which do not share a filesystem.  When the user runs
condor_submit, a command line option ("-s") can be given to tell
condor_submit to send all the files needed by the job to the schedd.
This is accomplished by first inserting the job into the queue
in the hold state.  Then, the schedd starts a new thread (or on
Unix, a new process) to receive all the files via the FileTransfer
object, and the files are stored in a subdirectory per job in the
spool directory.  Once all data files have been spooled, the
schedd modifies the job classad so all filenames are changed to
be relative to the job IWD, which is also modified to point
to the spool subdirectory.  Any attributes so modified are
"backed up" by prepending SUBMIT_ to the attribute name.  The
job is then released (moved to idle state).  The job is not removed from
the queue after completion until the output files are fetched, or
unless 10 days have passed.

This code is a first pass, and will undoubtedly undergo refinement
as time marches on.  But DZero folks need to start playing with it
immediately, so into CVS it goes.  Questions/concerns? See Todd T.