+ Previously, the environment variable CONDOR_IDS was only examined if the daemons were started as root.  Now, the variable is also examined if the daemons are started as a user with a regular uid. Normally, multiple users can submit to a schedd only if the realuid of that schedd is root or condor.  Now, muliple users can submit to a schedd even if that schedd is not running as root or condor, but only if the uid in CONDOR_IDS = the real uid of the daemon.
+ Think I fixed the following bug: if the daemons were started as root,
and the user "condor" did not exist, and CONDOR_IDS environment variable
was not set, we are supposed to bail out with an error message.  I think
that was not happening because of a bug treating a uid_t like a signed
int (on most platforms, it is unsigned).  Hopefully fixed.