Every time the schedd now gets a match from the negotiator:
a) ATTR_NUM_MATCHES in the job matched is incremented by one.

b) if ATTR_LAST_MATCH_LIST_LENGTH  is defined in the job ad to
be greater than zero, then the name attribute from the matched
target is stored in the job ad as
   LastMatchName0 = "some-startd-ad-name"
   LastMatchName1 = "some-other-startd-ad-name"
   ...
LastMatchName0 will hold the most recent match name.  The list will
be rotated with a max length defined by job attribute
ATTR_LAST_MATCH_LIST_LENGTH, which has a default value of zero
(i.e. don't keep this list in the job ad by default).  The idea
of this list is to allow the job to prefer to match with a different
resource (or the same) as in the past...  (requested by DZero).
Could also be used as part of the solution to the "black hole" problem.