+ for globus universe jobs, if a job is marked as removed but there is still a jobmanager contact string (i.e. we have not confirmed the job removed), startup a gridmanager to finish the job.  help to prevent jobs from sitting around in X state forever.
+ when releasing a job from hold, release job into the job state
specified by ATTR_JOB_STATUS_ON_RELEASE attribute, or into IDLE if
that attribute is not present.  Upon release, delete the attribute.

+ allow "condor_rm -f" to be performed not only on job in the removed
state, but also upon jobs that were put on hold after being in the
removed state.  [i.e. allow jobs which we failed to removed to be
removed by force].