Incorporated Todd's fix to periodic removal:
Notice the key is switch job_ad->LookupInteger() to be GetAttributeInt()
instead.  So before we got here, abortJobRaw changed ATTR_JOB_STATUS to be
REMOVED **inside of a transaction**.  GetAttributeInt() will first look up
the value in the pending transaction --- the LookupInteger will not see the
change until the transaction is committed.