Fixed a few minor bugs.
1. Changed the way we call LookupAccountName() so that we don't pass the
domain as the system parameter. This is wrong. To call this function
correctly, we have to concatenate the domain and the account and pass
that as the account name (known as the qualified account name). The
system parameter should be NULL, implying that we use the local system
to perform the lookup.

2. Changed a few dprintf()'s so that we print out domain\username,
and folded a few lines that were larger than 80 chars wide.