+ The AIX linker is oddball. Even though it garbage collects unused function references like other UNIX linkers, it MUST first resolve everything to know what it can collect. This means if you have a .o in a library which technically needs an externed variable which isn't supplied in the executable(and none of the .o functions are called), the AIX linker markes the extern as an unresolved symbol before it can garbage collect it. This is annoying.
+ To fix this, I added the concept of a "prelinker" which is a GNU
linker used to prelink the object files relating the internal gcc
libraries and what gcc compiles into a gigantic .o file. Then the AIX
linker takes this prelink.o file and links it against the system
libraries without having to perform this oddball garbage collection
pass because it has already happened(with respect to our code) via the
prelinker. Why you ask? Because as of gcc 3.3 and binutils 2.13.2 the
GNU ld could not produce executables which linked against the system
supplied libraries due to miscompatibility and could not be used.

+ Carefully defined flags to ensure the gcc internal c++ runtime is compiled
statically into the code, but not anything else. The rest of the libraries
like the libc.a library are shared.

+ Set up various things on and off for the clipped port, fixed the PVM
specification to pick the right architecture, set up various things to describe
the machine.