Patch to hopefully fix problems reported were very rarely, the child process forked by the schedd get stuck in an infinite loop when the user does "condor_submit -s".  Also similarly, hopefully will fix problems when the child process forked by the collector would very rarely get stuck in an infinite loop when COLLECTOR_QUERY_WORKERS > 0 in the config file.
The patch is to exit via a call to exec() when we are the child
created by a call to daemonCore Create_Thread, in order to avoid destructors
from being invoked.  Crazy, eh?  I think so as well.  But if you can think
of a better patch that doesn't involve tons of development/refactoring,
I'm all ears.  :).  See comment block in the code for more details.