Don't redefine exit() on DUX because it doesn't work (cuz dux libc.a include atexit() inside of exit.o, so we get multiple definitions).
This is just a temp fix to get v6.6.3 out the door.  For the next
release, after testing to make sure it work, I will rework the code
so the child after a Create_Thread just call atexit() to register
our exit via exec function.  Assuming that (a) atexit still works, and
(b) that atexit functions are called before c++ destructors are invoked,
it should be a much cleaner/simpler/portable way to get what we want.