- Added NEGOTIATOR_MATCHLIST_CACHING option to the config file. If True, then a sorted cache of machine ads that match jobs of a specific autocluster (and from the same user+schedd) is maintained -- thus matches can be made much faster if autoclustering is used.  The default value for this setting is False until more testing is performed.  NOTE: IT IS LIKELY THAT MATCHLIST CACHING, IN THE CURRENT IMPLEMENTATION, IS INCOMPATIBLE W/ RESOURCE ADS THAT SPECIFY ATTR_WANT_AD_REVAULATE (i.e. matching w/ an ad multiple times).
- Added NEGOTIATOR_SIMPLE_MATCHING config option; defaults to False. If
set to True, then the matchmaker doesn't get all ads from the collector;
instead, it just gets startd and submittor ads.  So if it is True,
the negotiation process may be speeded up on extremely large pools
(like Micron's) that have lots of masters/schedd, BUT if True, then
gangmatching-lite will not work (thus it defaults to false).

- Matching diagnostics switched from a set of bools to a set of
integer counters.