Fixed another bug in the firewall helper class. This time, I discovered that multiple calls to addTrusted() would cause a seg fault on Windows 2000. This has to do with how the code works if certain components of the windows firewall API aren't there, and Windows 2000 is the most exterme case, since NONE of it is there.
To address this, I have added an additional flag to the class that
indicates whether it is operational or not. The class is only flagged as
operational if the firewall API is available, and the firewall is turned
on. In all other case, the firewall class is considered not operational,
and all calls to its various functions do nothing and return false.