+ Added some prototypes for linux and solaris for getdents64().
+ Added a remap from getdents64() to getdirentries64() for linux because
getdents64() under linux is not a global symbol in the libc. So, since
this file links with the standard shadow, which needs to call this
function in receivers.o, I'll add it here. I don't like this at all,
but I don't have time to clean up the original problem and sloppy fix
and now I have to extend it. :(