+ removed some dead code.
+ Cleaned up the error messages when the shadow cannot open up the std files
and when it cannot chdir to the initial directory to include human readable
errors.

+ Fixed a bug where (before the job even ran) if the shadow failed to
chdir() to the initial directory or if it failed to open up a std file,
it would write a termination event into the user log (via a fall through
case which used many global variables which had not been altered away
from their default states since the job never actually ran) and then
except (due to said global variable defaults). This had the effect of
an execute followed by a terminate sequence of events in the user log
being repeated over and over.

Now, when this error condition happens, the shadow excepts right away
with a more verbose message about what went wrong and doesn't write the
erronous termination event.