+ Moved getdents64 into the solaris, irix, linux specification along with getdents so it can compile again on hpux 10.20.
+ Hpux doesn't really have a getdents64(), but it does have a getdents()
in the libc, but no prototype. I suppose people are to use getdirentries()
on hpux which internally uses getdents and we seem to catch that.

+ I don't know if as of this checkin, getdents() works on a remote i/o
situation or if it is used by readdir(). I might have to revisit the
implementation of getdents() in remote i/o under HPUX later when I know for
sure.