Backed out my 2005-02-24 change which removed support for the dag_node_name submit file command -- although the change was a good idea, and correctly implemented, it caused DAGMan to fail when used with older versions of condor_submit because of a condor_submit bug (gnats #447).
The condor_submit bug is that it cannot differentiate +FooBar=1 (i.e.,
attr-insertion) from foo_bar=1 (i.e., a submit file command), if the
latter is valid.  Although I intend to fix this, it won't help DAGMan
with an older, pre-fix condor_submit, so I still need to back out this
original change to condor_submit and DAGMan.

However, whenever it becomes safe for us to assume that a newer DAGMan
will never be used with a pre-6.7.7 condor_submit, support for the
dag_node_name submit file command should be re-removed, and DAGMan
should use +DAGNodeName="foo" instead...