added the plumbing to store and receive your parent's unique ID, which is needed for cleanup up sessions when reaping children.
a couple new attributes are placed into the security negotiation classad
and cached session.  these are pulled from the environment if daemoncore
did not set it for us, because there's no guarantee that we are linked
with daemoncore (*cough* old shadow :)

a new method, invalidateByParentAndPid, uses this information to clear out
sessions.