Fix a couple bugs if NEGOTIATOR_MATCHLIST_CACHING = True in the config file:
1) If you've got one user who submits a lot of jobs (where "lot" is >
number of resources) in one autocluster, and there is only one user in
the system, the MatchList will run out of machine ads...and stay out.
Future negotiation cycles won't re-build the MatchList because it's
always the same autocluster/schedd/etc.  The solution I chose was to
nuke the MatchList at the beginning of each negotiation cycle. Thanks
to Mike Yoder for catching this one.

2) Bad things could happen if (somehow) a negotiation cycle started
with no machine ads in the collector.