+ Added the interface to system() to execute an arbitrary program. In standalone mode, it acts just like normal system and is implemented how POSIX says it should be.
+ In remote mode, the environment of the user job is packaged up along
with the command and sent for execution by the shadow in terms of the
submit machine as the user.

+ The shadow has a right to deny the system() call as per the
administrator's wishes.

+ There is a bug where if there is whitespace in the environment, the shell
will attempt to execute things that it shouldn't out of the environment.
Eventually, we need to shell escape the various problem characters (probably
do this in the syscall library code).

+ The waitpid call which is used for solaris and linux have pretty different
implementations, under solaris it is extracted but under linux I found I could
call the system call directly.