Added support for discounting suspended claims at the matchmaker.
If we see a startd that is claimed but suspended, drop it from our list
of matches, as if the match has been broken. The negotiator won't reassign
it, but we won't charge the user for it. Similiarly, don't insert a startd
into a list of matches out on the pool if it's suspended.

Also, convert the two util functions to be member functions, we now refernece
a class variable to decide if we want this functionality. Todd swears that
someday there will be multiple instances of the Accountant object, so we're
going to have to pay the price someday.