First check-in of Nick Coleman's ClassAd analysis stuff into a release branch. Yeah!
This was originally on the trunk. It relies on new ClassAds as an external
though. Since we're not moving to the trunk anytime soon, we back-ported it
to Condor 6.7 so we can get its goodness.

Note that this directory has some experimental stuff of Nick's (the gangmatching
stuff) but it's not compiled by default, so it won't cause any problems.