Summary: The schedd's "DestroyProc()" function now returns specific error codes explaining what happened.  The Gridmanager uses this information to determine that trying to remove a job that doesn't exist is harmless.
Long version:

If the gridmanager tried to DestroyProc a job that it has already
successfully destroyed, it would abort the entire schedd
transaction.  This means that if a single job got confused and
DestroyProc got called twice than the gridmanager basically gave
up on destroying jobs (and possibly updating attributes).  Not
Good.

Problem 1 is "Why does it do this?"  The short answer is "Because
it was easy and believed to be harmless."  We were wrong.  The
same code path can cause multiple "005 Job Terminated" events to
be logged which confuses DAGMan, so this is clearly important.
Fixing this it on our list.

Problem 2 is "Is it a good idea to have everything fail because
we tried to Destroy a Proc that was already gone?"  Clearly not.
Thus, the change.