+ Added the 'want_remote_io = true | false' feature for the standard universe.
+ When the very first REMOTE_CONDOR_get_file_info(_new)? sender gets
called in the user job, the shadow will look into the job ad to see if
ATTR_WANT_REMOTE_IO is present and set to False. If so, this RPC call will
return with a value that shall be interpreted by the user job to mean that
all future access through the FileTable shall be 'local:' and to NEVER
ask the shadow again for how/where to open any future file.

+ The new return value of the above rpc call that signifies the use
of this feature by a newer shadow will be ignored by an older user job
without this new implementation.

+ Checkpoint files will still be honored by the shadow since that is the
first call back to the shadow and the shadow can still make a decision
about that particular file.

+ condor_submit knows about the 'want_remote_io' attribute and treats it as a
native attribute the user can set.

+ Once a job has decided to use this method of access, it holds true even
across checkpoint boundaries and the resuming job expects the execute machine
to have the same filesystem layout as the last time it ran.

+ I had to implement this functionality under duress. :) So I'm not quite sure
if there were any edge case side effects of this alteration.