Fixed a few code paths that kill the negotiator when a startd has a Name attribute with more than 63 characters in it. We assume that the Name can't be more than MAXHOSTNAMELEN characters, but this is totally broken. Any individual label in a name can't be more than that (example and com are both lables in example.com), but the entire name can be up to 255 characters. And even if DNS limits it, we shouldn't crash when it's too long.
This code was fixed by doing using an alternate version of LookupString
that mallocs the memory. Because we're not in the 6.7 branch, we can't use
the version that returns a MyString, which would be more elegant (we convert
to a MyString in a few places). Ideally some of this would be converted
to the MyString variety in the v6.7 branch.