Committing revised MANAGED behavior after review from Jaime.
Previously Managed was TRUE or FALSE.  TRUE meant an external force (usually condor_gridmanager) was controlling the job.  FALSE or undefined meant that the schedd was responsible.

Now Managed is a string.  "Schedd" means the schedd is managing it.  "External" means the gridmanager has it.  "ScheddDone" means that the external managing program is no longer managing the job _and no longer want it back_, essentially, the job is terminal in some form.

Beneficial result: The gridmanager no longer repeatedly takes control of jobs that are done.  This speeds up jobs leaving the queue, reduces the load of the gridmanager doing unnecessary work, and eliminates one particularlly common source of duplicate terminate events for jobs.