FileTransfer::DoUpload's interface changed.  It's now possible for Info.bytes to be non-zero, yet the transfer as a whole to fail.  This code was never updated to account for this.  Result: If the transfer failed, but at least 1 byte was transfered (or at least reported as transfered),   In the specific case of the shadow, if you have an X509 proxy, and that proxy transfer fails (Carey's proxy delegation code basically ensures that it will), the transfer will fail, *but the starter won't notice*.  The job exits as a success!   Files are silently not transfered.  BAD BAD BAD.
This is a partial fix.  Now FileTransfer::Upload notices the
error and propogates it up.

On the downside, it's only a partial fix. The higher levels seem
a bit confused.  In the case of the starter, it assumes it were
disconnected and hang around forever, claiming:

9/28 13:58:40 JIC::allJobsDone() failed, waiting for job lease to
expire or for a reconnect attempt

That case is as of yet unresolved.