+ This checkin changes the Requirements advertised in the machine ad from
Requirements = START

to

Requirements = (START) && (IsValidCheckpointPlatform)

+ It also adds a default expression for IsValidCheckpointPlatform which
is also published in the machine ad. IsValidCheckpointPlatform is an
epression which returns true for all non standard universe jobs, and true
under certain conditions for standard universe jobs.

+ The IsValidCheckpointPlatform expression can be redefined via the
IS_VALID_CHECKPOINT_PLATFORM in the config file. The default expression
says that any job which has never checkpointed before can start on this
machine--provided the machine exports a CheckpointPlatform attribute,
but any previously checkpointed job must match the checkpoint platform
of this machine (being there is a LastCheckpointPlatform in the job ad
set upon successful checkpoint). This default expression might change so
please see the actual code for details.

+ It was implemented in this manner both the matching of checkpoint
signatures is kind of out of band, can be very complex, and doesn't exist
for older versions of Condor. This manner provided us with the least confusing
option for specifying the matching expression and providing backwards
compatibility for older configuration files.

+ Fixed some buffer overflows with MyString.