+ Removed the condor_dh.o file from libcondorapi.{a,so}. This object file is not needed to use the functionality of libcondorapi.
+ Specifically it was removed because it had undefined references in it
with respect to standalone programs that link against libcondorapi.so.
We had never seen this problem before when there was only libcondorapi.a
because the linker threw away condor_dh.o (and the undefined references)
when noone referenced it, but when libcondorapi.so became available the text
segments were joined as one, forcing an unresolved linker error with functions
in condor_dh.o.