Add rounding of scalar job ClassAd attributes -- desirable so the autocluster code can "bucket" attributes like ImageSize, etc.
Add the following to the config file
   SCHEDD_ROUND_ATTR_<foo> = X
where <foo> is the job attribute to round, and X is the number of
decimals to round.  Rounding always rounds UP.  So to round a value up
to the nearest 100, set X = 2.  The actual, non-rounded value will
also be stored in the job ad as an attribute named <foo>_RAW.

The following are set by default:
   SCHEDD_ROUND_ATTR_ImageSize = 4
   SCHEDD_ROUND_ATTR_ExecutableSize = 4
   SCHEDD_ROUND_ATTR_DiskUsage = 4

Thus, the above attributes are all rounded up from kbytes to the nearest
10 megabytes.