Place the grid monitor files in a directory under the gridmanager scratch directory. When the gridmanager gives up on a grid monitor process, it deletes the directory.
Before, if the grid monitor was still alive, it could send another job
status file which wouldn't get cleaned up until the gridmanager exited.
Now that can't happen, as the directory it wants to write in doesn't
exist.

Another benefit is that the failed file transfers will cause the grid
monitor process to exit shortly after the gridmanager gives up on it.