Fix how we determine the Condor version of a Condor executable.
In CondorVersionInfo::get_version_from_file(),
we look for the Condor version magic string
'$CondorVersion: x.y.z <date> <extra info> $' in the executable file.
Before, we considered '$CondorVersion: \0' a valid match. This string
exists in all of our binaries, but comes after the real version string
on most platforms. On the platforms where the shorter string occurs
first (OSF and Windows), we get a bogus version. The most noticeable
result is that the grid universe doesn't work, because the schedd
misreads the version of the gridmanager.

Now, we look in the binary for a string that begins with
'$CondorVersion: ' and continues with a non-NULL character.