+ This checkin implements the "terminate pending" feature for the vanilla, java, mpi, parallel, and standard universes.
+ The point of this feature is to stop an
    execute, terminate, [execute, terminate]+
sequence in the log files. Once a terminate event has been written
then ONLY more terminate events can be written and the job must never
actually start up again. The behavior now is that there can be one or
more terminate events in the log files (and one of more termination
emails sent to the user).

+ The crux of this implementation is the realization that when a job is
considered to be completed, the shadow places a ATTR_TERMINATION_PENDING
= TRUE in the job ad which gets commited to the schedd. If the commit
to the schedd succeeds, then the job is in the "terminate pending" state
(except that it really isn't a true job state, just a contract the shadows
have with themselves to "do the right thing" when they see that attribute
set to true in the job ad upon invocation).

Then, if after the successful committing of the job ad, the shadow
excepts, segfaults, blows up, or who knows, and eventually another shadow
will be spawned to run the job again, and it'll notice the terminate
pending state and do a quick exit just writing a terminate event and
sending email, if applicable.

+ In this feature as implemented in this checkin, the schedd has
no knowledge of this behavior, and will happily match a job, gather
resources, and fire off a shadow only to have the shadow immediately
turn around, say it is finished, and the job then actually leaves the
queue. Later checkins can fixup the behaviour of the schedd to not be
so wasteful with resources.

+ The standard universe shadow's implementation is embarassing since the
code layout of that daemon is rife with heavily used global variables. I
basically stuck many of the global variables into the job ad itself for
easy reconstruction later during a terminate pending event. Some global
variables like JobStatus (gotten from waitpid's status) are actually
impossible to reconstruct manually--there are no analogues for WIFSIGNALED
to set those bits, so it had to go into the job ad.

+ The "new" shadow's implementation was easier to do, and involved having a
flavor of call for the functions that write the log termination events and
send the mail so the functions know what they might need to reconstruct
if in a pending state.

+ The local, scheduler, pvm, and various grid universes have not been adjusted
for this feature fix. That will probably happen in the near future.