Fix a problem that happened when file transfer fails to return output files on job exit.  When this happened, we couldn't tell why the file transfer failed -- we assumed that the command socket or the network had died, and we waited for a reconnect.  In some cases, this wasn't true, the socket was intact, and we would wait forever for a reconnect that would never come.
So, now, when this happens, we force the command socket to close, which will
make the shadow reconnect to us.  But we only do it five times just in case
there's some other error, after which we really give up.

Ugh.  The real solution is to have better error propogation from the
file_transfer object to the starter and back to the shadow.