Fixed multiple bugs in CEDAR connect code, primarily related to non-blocking connections.  The worst resulted in the registered callback never happening, because the socket got into an unexpected state (sock_bound) upon return from Sock::do_connect_finish().  The resulting buildup of registered sockets could eventually cause the program to run out of file descriptors.
Also improved error reporting and handling for both non-blocking and
blocking connections.  ECONNREFUSED errors are now detected in the
case where connect() returns EWOULDBLOCK.  Receipt of a signal within
CEDAR's connect select loop no longer causes the connection attempt to
fail.  Moved code from do_connect_tryit() to cancel_connect(), and
ensured that is is called on all code paths where a connection attempt
has failed--not just the case where connect returns something other
than EWOULDBLOCK.  Also added code in cancel_connect() to call
Sock::timeout() if it closes and reopens the socket, since not
doing this could cause potentially very long blocking on the
next call to connect.

The meaning of connect_state.timeout_interval was not consistent
between blocking and non-blocking connections.  For blocking
connections, this was the timeout after which we stop making more
calls to connect().  For non-blocking, it was the timeout of the call
to connect().  Fixed the non-blocking case to have the same behavior
as the blocking case.  It uses Sock::_timeout as the connect timeout
and it uses connect_state.timeout_interval (now renamed to
retry_timeout_interval) as the timeout after which we stop making new
attempts to connect.

Connection retries for non-blocking connections no longer block for a
second in sleep.  Control returns to the main select loop, which then
calls do_connect_retry() after a second.