+ This checking implements the "master kills orphans from previous instantiation of Condor" feature.
+ The procInfo/Raw structure now has the command line of the process available.
It returns real values on linux and solaris, and a placeholder value
for all other platforms for later implementation.

+ The procInfo/Raw structures now have a constructure and destructor to deal
with the memory associated with the command line.

+ The master now params for KILL_ORPHANS_ON_STARTUP, which defaults to true,
and if true, then kills the orphans.

+ The pidenvid ancestor variables have been extended to include the birthdate
of the forker process on the LHS.

+ The pidenvid library can now find the oldest ancestor out of a PidEnvID
structure. Another function was added to copy individual PidEnvID entries.

+ Added ProcAPI::killOrphans(void) to the procapi library. Read the
extensive comments in that function for how and why it does what it does.