+ This commit unchecks in the Orphan killing feature of the master.
+ Why? (excerpted from a YIM conversation)

- So, in NMI, we start up a personal 6.8.2 Condor under a 6.7.20 condor
master (which is the NMI system). I had to extend the size and alter the
format of a bunch of static arrays to hold the environment ancestral
data--the PidEnvdID is a staticaly sized structure due to memory
management complexities the original version had to deal with. So,
when the older daemon read the newer stuff, it would overwrite memory,
and when the new daemon read the old stuff, it would kill pidenvid enabled
process on the machine in an undefined manner. So it was really backwards
compatibility and a strcpy instead of a strncpy which doomed it in it
current implementation.

- After I check in the undo, I have to fix the original security problem,
check that in, and then probably have to really rethink how to keep the
ancestral environment information both in the daemon's memory and in the
environment of the process itself. Basically, I have to be able to read
ANY revision of the data out of a process, which, through mixed condor
version trees, could actually have multiple versions of the ancestral
variables in an environment.