Add new classad attributes PrivateNetworkIpAddr, PrivateNetworkName, and PublicNetworkIpAddr. They are the new official way for a daemon to advertise its address in its daemon ad.
If GCB is not being used, PublicNetworkIpAddr is the daemon's real
address. If GCB is being used, PublicNetworkIpAddr is the daemon's proxy
address on the broker, PrivateNetworkIpAddr is the daemon's real address,
and PrivateNetworkName is the name of the daemon's private network
(defined by PRIVATE_NETWORK_NAME in the config file).

The Daemon object now looks for these attributes when determining the
address of a daemon. If they are not present, it falls back to the old
address attributes.