Better error message when PREEN_INTERVAL is set to 0.
Historically PREEN_INTERVAL=0 was interpreted as "use the
default," which was surprising.  So the recent change that forbad
setting it to 0 was a good idea.  Unfortunately at least one user
erroneously believed that PREEN_INTERVAL=0 would disable
condor_preen.  Since there may be other users with similar
erroneous beliefs, so now the code generates a slightly more
helpful error message, specifically noting that to disable
condor_preen you need to set PREEN="".

(For the version history checker: this will go in the version
history in a more general statement about multiple settings no
longer accepting 0 to mean "default."  That's what is user
visible.)