+ Fixed the "bug" where the checkpoint server didn't honor HIGHPORT and LOWPORT. It was causing some headache for INFN and someone on condor-users. It appears their user jobs were getting assigned to a checkpoint server and that the firewalls wouldn't let the jobs talk to them since the checkpoint server allocated ports outside of the allowed ports for the pool.
+ The well-known ports for the checkpoint server are still well known,
but the child processes which do the store/restore to the actual job
itself are now restricted to LOWPORT and HIGHPORT.

+ Also fixed a bug where we hotns a port number, but then later check it,
_in that form_, against 1024 and change to root if less than 1024. This
means were were granting root provledges for ports we didn't have to.

+ Added some better debugging information. Now the checkpoint server writes
out what ip:port it is giving to the shadow on which the job should connect.
This is much more useful for debugging firewall problems of this kind.