Be more careful about how we decide where a file came from - it's not always just the IWD of the job.
At the start, we know the basename of the file we're being sent. So,
put together a list of all possible input files (TransferInputFiles, Input,
Exec, etc) and find a matching basename. Take the dirname of that - if it's
absolute, use it. If it's relative, stick it on to Iwd. If it's '.', just
use Iwd.

ToDo: add X509s that we're delegating, not transfering. It's ok to miss them
for now.