Upon a disconnected shadow->starter reconnect, the public address of the schedd is passed along to the startd.  This is done because if the schedd restarted, it probably did so on a different port, and the startd may need to know the new address in order to keep sending keepalives (if the claim keepalives are being sent from the startd to the schedd).
It works as follows: the public addr of the schedd is passed to the
shadow via argv.  the shadow sends it to the startd when it sends the
request to locate the starter (backwards compatible because the
communication is via classad).  the startd then stores the new schedd
address in the correct Client object, where it is used by the keepalive
code.