Remove unnecessary checks for the condor uid when trying to switch to the user uid.
This was precipitated by a change to start the gridmanager as root with
effective uid of the user. The default gridmanager log includes
$(USERNAME). To have $(USERNAME) continue to evaulate to the user rather
than 'root', the gridmanager changes its effective uid to the user in
main_pre_dc_init(). This resulted in a call to init_condor_ids(), which
would fail if the system didn't have a condor user. If the admin set
CONDOR_IDS in the config file, this shouldn't happen, but
main_pre_dc_init() executes before the config file has been read.

The comment in _set_priv saying that init_condor_ids() initializes
SwitchIds is out of date. can_switch_ids() does that the first time it's
invoked. Thus, init_condor_ids() doesn't have to be called in
_set_priv().