Fixed EXCEPT in Directory class
This commit fixes a variety of cases that can lead to EXCEPTs
happening in the Directory class. The one that has been seen
in practice happens in the Starter when calculating a job's
disk usage if the job happens to be a glide-in using the
GLEXEC_STARTER feature with GLEXEC_TMP_DIR set somewhere in
the sandbox.