Removed condor_amazon, should not be here yet
The condor_amazon directory is one created on the V7_1-amazon_ec2
branches. It should not be in the main V7_1 development branches
yet. It will be created when the V7_1-amazon_ec2-2-branch is merged
into V7_1. It existed on the main V7_1 development branch, aka master,
because of the conversion from CVS->SVN->Git.