Removed FORTRAN_HAS_RECURSION and the requirement of recursion for f77 code.
There is now a new file in condor_tests called fort-gen.pl. It is not currently
used, but it is the program which generated job_ckpt_f-float.f and
job_ckpt_f-integers.f. The generator has a comment about how to invoke it.

The generated programs now do not exhibit recursive calls, but instead one long
generated call stack if individual function names. The output of the generated
programs was made more stable by adding format statements.

The .run file associated with each test has been altered to expect the new
output the programs create.

It should be that the build system _always_ generates the fortran code when
make all is done, but that is a future feature.