Derek's startd fixes for V7_1_0. Dan is committing this to merge in Derek's startd fixes for V7_1_0.  These changes were made on the abandoned CVS repository, as planned. Here are Derek's CVS commit messages for the commits bundled together here:
Made the startd smarter about not churning through the work-fetch
code if not configured that way. There are two main changes here:
- If the FetchWorkDelay expression is undefined, default to 5 minutes
instead of not using a delay at all.
- If the slot doesn't have a hook keyword, don't bother considering
anything else (in-progress fetches, the delay expression, etc).

Added bogus return value to Resource::reconfig() to avoid a compiler
warning about control reaching the end of a non-void function. The
return value is just to make the prototype happy for ResMgr::walk().

Moved control of the hook keyword from the StartdHookMgr object into
the Resource object for each slot. Not only is the code to manage te
strings cleaner this way, but it lets the Resource object directly
know if it is configured for job hooks or not. This commit also
includes a change so that the Resource has a reconfig() method which
is called on every Resource by the ResMgr during a reconfig.

Renamed HOOK_REPLY_CLAIM to HOOK_REPLY_FETCH to be more clear about
what it is and when it's invoked. This hook is called every time the
startd fetches work to let the outside world know what happened to the
work it fetched. The hook no longer has anything to do with claims,
now that claims can be reused for fetched work.