Initial revision of the splice feature set.
A splice is another dag which is brought in and assigned a name. This name
can then be used in PARENT/CHILD relationships. All initial or final nodes
in a dag are treated as a set for dependency purposes.

Also implemented a new feature were $(JOB) is interpolated with the correct
value in the value of a vars line for a job.