Improve timing in some standard universe tests
Some of the standard universe tests need to spin for a certain number of
seconds. We use to do this by running a calibration program to see how
many iterations of a tight for-loop the machine could chew through in a
second. But we've seen cases where load on the machine causes the
calculation to be off by an order of magnitude. One time, a test that
was supposed to run for 45 seconds finished in 5, which didn't leave
enough time to force a checkpoint.

Now, we call time() or gettimeofday() periodically in a loop to
determine when the correct amount of time has passed. This adds a new
dependency on the syscall librarie's gettimeofday(), which goes back to
the shadow on the first call, but it's an improvement.