Change behavior of names ending in an @ sign to allow for multiple Schedds to report the same SCHEDD_NAME. This is critical for transparent Schedd fail-over.
Issues is SCHEDD_NAME always has a FQDN appended to it, making it impossible to do transparent fail-over.

A quick look at the code suggests that the place all this happens in src/condor_c++_util/get_daemon_name.C, which has get_daemon_name and build_valid_daemon_name.

build_valid_daemon_name is called from daemons when they init to create a name to advertise to the Collector. get_daemon_name is used by tools to parse names provided by -name arguments before passing them to daemon clients.

build_valid_daemon_name:
// Given some name, create a valid name for ourself with our full
// hostname. If the name contains an '@', strip off everything after
// it and append my_full_hostname(). If there's no '@', try to
// resolve what we have and see if it's my_full_hostname. If so, use
// it, otherwise, use name@my_full_hostname(). We return the answer
// in a string which should be deallocated w/ delete [].

The consequence of this behavior is that it is impossible to define a name that does not contain a fqdn.

get_daemon_name:
// Return a pointer to a newly allocated string that contains the
// valid daemon name that corresponds with the given name. Basically,
// see if there's an '@'. If so, resolve everything after it as a
// hostname. If not, resolve what we were passed as a hostname.
// The string is allocated with strnewp() (or it's equivalent), so you
// should deallocate it with delete [].

get_daemon_name also returns NULL if the given name has no '@' and does not resolve with get_full_hostname, or if there is an '@' and the data after it does not resolve with get_full_hostname. NULL causes tools to exit with an error that they could not lookup the name.

Digging deeper suggests that Daemon.C works hard to parse the given -name, but only succeeds in very specific cases, which I was not able to generate. When the tools fails to parse the -name it always falls back to looking the name up in the Collector.

This patch changes the behavior of what happens when a name ends with an @. Currently, my_full_hostname will be stuck on the end of it. This patch just leaves the name alone. The result is being able to have a SCHEDD_NAME=foo@ and do condor_q -name foo@. This seems to be the patch of least change since the foo@ case should just be a corner case.

A deeper solution may be to investigate how hostnames are being used and consider removing them in places, which may also reduce Condor's reliance on DNS. For instance, 'condor_q -name foo' will fail if foo does not resolve in DNS, which seems simply wrong since the name really just needs to resolve in the Collector.

Of course, names still need to be fairly unique because of how the Collector stores information, but in this specific case uniqueness is actually blocking a proper feature.